requirements.txt
rdflib>=4.2.2
python-dateutil~=2.8.2
boto3==1.11.7
geographiclib==1.50
requests~=2.26.0
setuptools~=56.0.0
botocore~=1.14.17
python-dotenv~=0.19.2



